# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/backstage/role": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-Role"
                ],
                "summary": "Role View",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "int default",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            15,
                            20,
                            30,
                            40,
                            50
                        ],
                        "type": "integer",
                        "description": "int enums",
                        "name": "pageLimit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "string enums",
                        "name": "sort",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "id",
                            "key"
                        ],
                        "type": "string",
                        "description": "string enums",
                        "name": "sortColumn",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "",
                        "description": "string default",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "",
                        "description": "string default",
                        "name": "searchCategory",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backstagedto.RoleListDTO"
                        }
                    }
                }
            }
        },
        "/backstage/role/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-Role"
                ],
                "summary": "Role Create",
                "parameters": [
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backstagedto.RoleCreateOrEditDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/backstage/role/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-Role"
                ],
                "summary": "Role Delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/backstage/role/edit/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-Role"
                ],
                "summary": "Role Edit",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backstagedto.RoleCreateOrEditDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/backstage/role/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-Role"
                ],
                "summary": "Role By Id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backstagedto.RoleIdDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "backstagedto.RoleCreateOrEditDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "select": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "status": {
                    "type": "boolean"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
        "backstagedto.RoleIdDTO": {
            "type": "object",
            "properties": {
                "roleById": {
                    "$ref": "#/definitions/backstagedto.RoleCreateOrEditDTO"
                }
            }
        },
        "backstagedto.RoleListDTO": {
            "type": "object",
            "properties": {
                "pageData": {
                    "$ref": "#/definitions/dto.PageForMultSearchDTO"
                },
                "roleList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backstagedto.RoleViewData"
                    }
                }
            }
        },
        "backstagedto.RoleViewData": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "updateTime": {
                    "type": "string"
                },
                "updateUser": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
...
        "dto.PageForMultSearchDTO": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "pageLimit": {
                    "type": "integer"
                },
                "search": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "sort": {
                    "type": "string"
                },
                "sortColumn": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}